rename directory; move paths.js to utils
