+++ src/public/js/components/baseView.js
