+++ src/public/js/utils/paths.js
@@ -1,4 +1,4 @@
-export const paths = {
+export const PATHS = {
     login: '/login',
     signup: '/signup',
     profile: '/', // временно, пока не готовы главная страница и остальные контроллеры
